feat:

+++ src/main/java/com/upravdom/oko/repositoriy/CityRepository.java
@@ -1,7 +1,7 @@
 package com.upravdom.oko.repositoriy;
 
-import com.upravdom.oko.model.City;
+import com.upravdom.oko.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
 }
